Separate CCI subdirectories into libraries.
Each subdirectory found inside lib/ is now a library that
can be linked to others, as well as tools found inside src/.
Names have the `CCI` prefix and a upper-case first character.
For instance, lib/lex is CCILex, lib/parser is CCIParser and so on.

This eases the process of unit testing, now that integrating with
one of the CCI parts becomes as trivial as linking each library
to the final unit test executable.