nit: prepend instead of append to CMAKE_MODULE_PATH
The reasoning behind this is that,
I usually want to make sure the project local module is picked up,
and not use the CMake-provided one in case the API differs.

One situation where I would want to append,
is I was backporting a CMake feature in the project.
I would want to pick up the updated version from CMake if available,
but otherwise fall back to the in-project version.