fix compilation with GCC 7.2.1 on Linux
nit: prepend instead of append to CMAKE_MODULE_PATH

The reasoning behind this is that,
I usually want to make sure the project local module is picked up,
and not use the CMake-provided one in case the API differs.

One situation where I would want to append,
is I was backporting a CMake feature in the project.
I would want to pick up the updated version from CMake if available,
but otherwise fall back to the in-project version.

use standard BUILD_TESTING instead of CCI_ENABLE_TESTS

and give instruction how to run tests with CTest

remove INTERPROCEDURAL_OPTIMIZATION from CMakeLists.txt

Leave this decision up to a toolchain file.

The idea, I have, is that the CMakeLists.txt should contain the necessary
information to build the targets.
But the tuning, which can be different between compilers,
is better handled by a toolchain file,
it's less intrusive, and the project files won't look compiler-specific.

cci_lex: fix cci_basic visibility PRIVATE => PUBLIC

include/cci/lex/lexer.hpp includes cci/basic/source_manager.hpp,
so cci_basic is part of the public interface of cci_lex.

unittest: test contract only if CCI_ENABLE_CONTRACTS is defined

fix GCC compilation issue when CCI_ENABLE_CONTRACTS is not set

GCC gave the following error:

    ../include/cci/util/contracts.hpp:51:39: error: missing binary operator before token "("
     #if defined(__GNUC__) || __has_builtin(__builtin_unreachable)
                                           ^

This change fixes that, by defining __has_builtin
for compilers that don't provide it.

Also, fixed cci_unreachable(cond) => cci_unreachable().

And simplify the condition:

    #if defined(__GNUC__) || defined(__clang__)

by:

    #if defined(__GNUC__) || defined(__clang__)

The reason being that Clang defines __GNUC__.

make CCI_ENABLE_CONTRACTS target local (added cci_util target)

Make sure CCI_ENABLE_CONTRACTS is always defined, to 0 or 1.
I believe this is a good practice.
This allow to detect with GCC's -Wundef that the flags is always handled,
and avoids mistake in the build dependencies.
This allowed me to detect that the unittest did not depend on cci_util,
when it was testing the contract.

Another possibility, would be to make a config.h from a config.h.in,
with configure_file().
This config.h.in would contain the line:

    #cmakedefine01 CCI_ENABLE_CONTRACTS

And contracts.hpp would include config.h.

lexer: write switch-case so that uncovered cases can trigger a warning

At least with GCC 7.2, if you put a default case,
the compiler estimates that the switch does not need to be fully covered,
and does not trigger a warning.

Removing the default case, handling that after the switch reveals a warning,
e.g.:

    ../lib/lex/lexer.cpp: In function ‘std::string_view cci::lex::to_string(cci::lex::TokenKind)’:
    ../lib/lex/lexer.cpp:212:10: warning: enumeration value ‘kw_continue2’ not handled in switch [-Wswitch]
       switch (K)
              ^

remove fmt deps

As it is now expected to be installed by the user,
it does not look necessary to bundle the code with the project.

add return statement to main

make git_revision.cpp self-contained

unittest: remove obsolete file

unittest: remove explicit GTEST_INCLUDE_DIRS

target_link_libraries(... GTest::GTest ...) is sufficient,
it exports the directories.

Also removed -Wno-zero-as-null-pointer-constant.
This is a toolchain issue, here the code would not compile on some platforms,
e.g. Windows, because the flag is not supported.

The best is to put -Wno-zero-as-null-pointer-constant in a toolchain file,
with other warnings.

By the way, it looks like this is an issue with Clang <= 5.0,
see https://bugs.llvm.org/show_bug.cgi?id=33771

So the best way to handle this, might be to use -Wzero-as-null-pointer-constant
only after Clang 6 comes out.

clang toolchain: add warnings and clang tools

I cherry-picked some of the flags that could be found in the previous
version of your CMake files, feel free to adjust,
the goal is just to give an idea of what can go in a toolchain file.

The nice thing, with toolchains, is that you can experiments with new compiler
and stuff, but the project does need to be modified.

Often when you do this in a project directly,
you introduce incompatibility with compilers you don't use.

The libc++ stuff has not been tested as I don't have it installed on my box,
it may require some adjustments.

Fix git_revision in CMake

Add note about clang toolchain in CMake