remove INTERPROCEDURAL_OPTIMIZATION from CMakeLists.txt
Leave this decision up to a toolchain file.

The idea, I have, is that the CMakeLists.txt should contain the necessary
information to build the targets.
But the tuning, which can be different between compilers,
is better handled by a toolchain file,
it's less intrusive, and the project files won't look compiler-specific.