make CCI_ENABLE_CONTRACTS target local (added cci_util target)
Make sure CCI_ENABLE_CONTRACTS is always defined, to 0 or 1.
I believe this is a good practice.
This allow to detect with GCC's -Wundef that the flags is always handled,
and avoids mistake in the build dependencies.
This allowed me to detect that the unittest did not depend on cci_util,
when it was testing the contract.

Another possibility, would be to make a config.h from a config.h.in,
with configure_file().
This config.h.in would contain the line:

    #cmakedefine01 CCI_ENABLE_CONTRACTS

And contracts.hpp would include config.h.