unittest: remove explicit GTEST_INCLUDE_DIRS
target_link_libraries(... GTest::GTest ...) is sufficient,
it exports the directories.

Also removed -Wno-zero-as-null-pointer-constant.
This is a toolchain issue, here the code would not compile on some platforms,
e.g. Windows, because the flag is not supported.

The best is to put -Wno-zero-as-null-pointer-constant in a toolchain file,
with other warnings.

By the way, it looks like this is an issue with Clang <= 5.0,
see https://bugs.llvm.org/show_bug.cgi?id=33771

So the best way to handle this, might be to use -Wzero-as-null-pointer-constant
only after Clang 6 comes out.