clang toolchain: add warnings and clang tools
I cherry-picked some of the flags that could be found in the previous
version of your CMake files, feel free to adjust,
the goal is just to give an idea of what can go in a toolchain file.

The nice thing, with toolchains, is that you can experiments with new compiler
and stuff, but the project does need to be modified.

Often when you do this in a project directly,
you introduce incompatibility with compilers you don't use.

The libc++ stuff has not been tested as I don't have it installed on my box,
it may require some adjustments.