Introduce cci::selector for diagnostic messages
`cci::selector` allows for a selection of a fixed set of options/cases
in a diagnostic message when reporting it. The new format is:

    message [ '|' message ]...

Rather than letting the caller choose a value for the format (e.g.
diagnosing an invalid digit for a numeric constant, which could be a
floating point, integer constant, be it octal, decimal, or hexadecimal),
the approach now is to encode all possible options in the diagnostic
message directly. For example:

    DIAG(err_invalid_digit,
         Severity::Error,
         "invalid digit for {0:floating point|integer} constant")

Reporting this error code is as simple as:

    diag.report(nocontext, diag::err_invalid_digit, selector{1});

Which will output "invalid digit for integer constant".